MovieObserver/backend/requirements.txt
# Web scraping libraries
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
lxml==4.9.3

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Database
pymongo==4.6.0
motor==3.3.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.1

# Data processing
pandas==2.1.1
